Trivial: move TextContextMenuItem into an independent file
